refactor: reduce complexity in 6 packages, fix topology dedup (ISS-284)
Removes all 33 fitness exceptions for syn-collector, syn-perf,
openclaw-plugin, syn-pulse-ui, syn-tokens, and syn-shared by splitting
large modules into focused files.

Also fixes a bug in the topology analyzer (aps-cli v0.1.1) where
tree-sitter's overlapping query patterns caused functions to be counted
2-3x, inflating file-level cognitive complexity and LOC metrics by up
to 3x. The fix deduplicates by (file_path, start_line) in
write_topology_artifacts() and corrects the per-module language list.
Net effect: measured violations drop from 591 to 265.

New files (29):
  syn-collector: routes.py, watcher_runner.py, event_id.py, parsing.py,
    hook_parser.py, transcript_parser.py
  syn-perf: compare_result.py, tables.py, comparison_report.py,
    parallel_report.py
  openclaw-plugin: format.ts, observability_session.ts,
    execution_detail.ts, artifact_detail.ts
  syn-pulse-ui: HeatmapTooltip.tsx, SummaryStats.tsx, YearSelector.tsx,
    HeatmapLegend.tsx
  syn-tokens: pricing.py, usage_summary.py, tenant_ops.py,
    redis_token_store.py, budget_stores.py, token_stores.py,
    singletons.py
  syn-shared: op_client.py, env_file.py, git_identity.py,
    git_identity_resolver.py, workspace_security.py

Closes #284
