+++ .topology/manifest.toml
@@ -1,13 +1,13 @@
 [topology]
 version = "0.1.0"
-generated_at = "2025-12-17T18:05:21Z"
+generated_at = "2025-12-17T23:21:01Z"
 generator = "aps-cli"
 generator_version = "0.1.0"
 
 [analysis]
 root = "."
-languages = ["tsx", "typescript", "python", "rust"]
-total_files = 1197
-total_functions = 10208
-total_modules = 855
-total_dependencies = 4712
+languages = ["rust", "python", "typescript", "tsx"]
+total_files = 1200
+total_functions = 6053
+total_modules = 863
+total_dependencies = 4660
