+++ packages/syn-collector/src/syn_collector/collector/service.py
@@ -8,12 +8,12 @@ from __future__ import annotations
 
 import logging
 from contextlib import asynccontextmanager
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 from fastapi import FastAPI
 
 from syn_collector.collector.dedup import DeduplicationFilter
-from syn_collector.events.types import BatchResponse, EventBatch
+from syn_collector.collector.routes import register_routes
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator
@@ -58,69 +58,6 @@ def create_app(
         lifespan=lifespan,
     )
 
-    @application.post("/events", response_model=BatchResponse)
-    async def receive_events(batch: EventBatch) -> BatchResponse:
-        """Receive a batch of events from a sidecar.
-
-        Events are deduplicated by event_id and written to
-        the observability store.
-        """
-        accepted = 0
-        duplicates = 0
-
-        for event in batch.events:
-            if dedup.is_duplicate(event.event_id):
-                duplicates += 1
-                continue
-
-            try:
-                await store.write_event(event)
-                accepted += 1
-            except Exception as e:
-                logger.error(
-                    f"Failed to write event: {e}",
-                    extra={
-                        "event_id": event.event_id,
-                        "batch_id": batch.batch_id,
-                        "error": str(e),
-                    },
-                )
-                continue
-
-        logger.info(
-            f"Processed batch {batch.batch_id}: {accepted} accepted, {duplicates} duplicates",
-            extra={
-                "batch_id": batch.batch_id,
-                "agent_id": batch.agent_id,
-                "accepted": accepted,
-                "duplicates": duplicates,
-                "total": len(batch.events),
-            },
-        )
-
-        return BatchResponse(
-            accepted=accepted,
-            duplicates=duplicates,
-            batch_id=batch.batch_id,
-        )
-
-    @application.get("/health")
-    async def health() -> dict[str, str]:
-        """Health check endpoint."""
-        return {"status": "healthy"}
-
-    @application.get("/stats")
-    async def stats() -> dict[str, Any]:
-        """Get collector statistics."""
-        return {
-            "dedup": dedup.stats,
-            "hit_rate": dedup.hit_rate(),
-        }
-
-    @application.post("/reset")
-    async def reset() -> dict[str, str]:
-        """Reset collector state (for testing)."""
-        dedup.clear()
-        return {"status": "reset"}
+    register_routes(application, store, dedup)
 
     return application
