+++ packages/syn-perf/src/syn_perf/benchmarks/throughput.py
@@ -49,13 +49,12 @@ class ThroughputBenchmark(BaseBenchmark):
 
         start_time = time.perf_counter()
         end_time = start_time + duration
-        active_tasks: set[asyncio.Task] = set()
+        active_tasks: set[asyncio.Task[Any]] = set()
         completed_count = 0
 
         self.log(f"Running throughput test for {duration}s with {max_concurrent} concurrent...")
 
         while time.perf_counter() < end_time:
-            # Start new workspaces up to max_concurrent
             while len(active_tasks) < max_concurrent and time.perf_counter() < end_time:
                 workspace_id = f"throughput-{uuid.uuid4().hex[:8]}"
                 task = asyncio.create_task(self.time_workspace_cycle(workspace_id))
@@ -64,43 +63,50 @@ class ThroughputBenchmark(BaseBenchmark):
             if not active_tasks:
                 break
 
-            # Wait for any task to complete
             done, active_tasks = await asyncio.wait(
                 active_tasks,
                 timeout=0.1,
                 return_when=asyncio.FIRST_COMPLETED,
             )
+            completed_count += self._collect_task_results(done, result)
 
-            # Process completed tasks
-            for task in done:
-                try:
-                    timing = task.result()
-                    result.add_timing(timing)
-                    completed_count += 1
-                except Exception as e:
-                    result.errors.append(str(e))
-
-        # Wait for remaining tasks
         if active_tasks:
             self.log(f"Waiting for {len(active_tasks)} remaining tasks...")
             done, _ = await asyncio.wait(active_tasks, timeout=30)
-            for task in done:
-                try:
-                    timing = task.result()
-                    result.add_timing(timing)
-                    completed_count += 1
-                except Exception as e:
-                    result.errors.append(str(e))
+            completed_count += self._collect_task_results(done, result)
 
+        self._finalize(result, completed_count, start_time, max_concurrent)
+        return result
+
+    @staticmethod
+    def _collect_task_results(
+        done: set[asyncio.Task[Any]],
+        result: BenchmarkResult,
+    ) -> int:
+        """Harvest completed tasks into *result*, returning success count."""
+        count = 0
+        for task in done:
+            try:
+                result.add_timing(task.result())
+                count += 1
+            except Exception as e:
+                result.errors.append(str(e))
+        return count
+
+    @staticmethod
+    def _finalize(
+        result: BenchmarkResult,
+        completed_count: int,
+        start_time: float,
+        max_concurrent: int,
+    ) -> None:
+        """Compute throughput metrics and mark *result* complete."""
         actual_duration = time.perf_counter() - start_time
         result.iterations = completed_count
         result.duration_seconds = actual_duration
         result.workspaces_per_second = (
             completed_count / actual_duration if actual_duration > 0 else 0
         )
-
         result.metadata["max_concurrent"] = max_concurrent
         result.metadata["workspaces_per_minute"] = result.workspaces_per_second * 60
-
         result.complete()
-        return result
