+++ packages/syn-shared/src/syn_shared/settings/env_file.py
@@ -1,0 +1,40 @@
+"""Environment file parser.
+
+Extracted from op_resolver.py to reduce module complexity.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from pathlib import Path
+
+
+def parse_env_file(path: Path) -> dict[str, str]:
+    """Parse a .env file into a key-value dict.
+
+    Lines starting with ``#`` and blank lines are ignored.
+    Values may be quoted with single or double quotes.
+    """
+    result: dict[str, str] = {}
+    try:
+        text = path.read_text(encoding="utf-8")
+    except OSError:
+        return result
+
+    for raw_line in text.splitlines():
+        line = raw_line.strip()
+        if not line or line.startswith("#"):
+            continue
+        if "=" not in line:
+            continue
+        key, _, raw_value = line.partition("=")
+        key = key.strip()
+        value = raw_value.strip()
+        if len(value) >= 2 and value[0] == value[-1] and value[0] in ('"', "'"):
+            value = value[1:-1]
+        if key:
+            result[key] = value
+
+    return result
